LayoutTests/platform/ios-simulator-wk2/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-self-baseline-not-applied-if-sizing-cyclic-dependency-002-expected.txt

auto-sized rows - horizonal grid and verticalLR item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 min-content-sized rows - horizonal grid and verticalLR item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-content-sized rows - horizonal grid and verticalLR item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 fit-content-sized rows - horizonal grid and verticalLR item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 flex-sized rows - horizonal grid and verticalLR item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-flex-sized rows - horizonal grid and verticalLR item - column-axis baseline
baseline is not applied initially, but orthogonal items force repeating the track sizing and height is not indefinite in that phase.
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 auto-sized columns - horizontal grid item - row-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 min-content-sized columns - horizontal grid item - row-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-content-sized columns - horizontal grid item - row-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 fit-content-sized columns - horizontal grid item - row-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 flex-sized columns - horizontal grid item - row-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-flex-sized columns - horizontal grid item - row-axis baseline
baseline is not applied initially, but orthogonal items force repeating the track sizing and height is not indefinite in that phase.
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 auto-sized rows - verticalLR grid and horizontal item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 min-content-sized rows - verticalLR grid and horizontal item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-content-sized rows - verticalLR grid and horizontal item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 fit-content-sized rows - verticalLR grid and horizontal item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 flex-sized rows - verticalLR grid and horizontal item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-flex-sized rows - verticalLR grid and horizontal item - column-axis baseline
baseline is not applied initially, but orthogonal items force repeating the track sizing and height is not indefinite in that phase.
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 auto-sized rows - verticalRL grid and horizontal item - column-axis baseline
baseline alignment is not supported, because the RL items do not share an alignment context with the LR items
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 min-content-sized rows - verticalRL grid and horizontal item - column-axis baseline
baseline alignment is not supported, because the RL items do not share an alignment context with the LR items
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-content-sized rows - verticalRL grid and horizontal item - column-axis baseline
baseline alignment is not supported, because the RL items do not share an alignment context with the LR items
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 fit-content-sized rows - verticalRL grid and horizontal item - column-axis baseline
baseline alignment is not supported, because the RL items do not share an alignment context with the LR items
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 flex-sized rows - verticalRL grid and horizontal item - column-axis baseline
baseline alignment is not supported, because the RL items do not share an alignment context with the LR items
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-flex-sized rows - verticalRL grid and horizontal item - column-axis baseline
baseline alignment is not supported, because the RL items do not share an alignment context with the LR items
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É

FAIL .inline-grid 1 assert_equals:
<div class="inline-grid alignItemsBaseline columns height200">
  <div class="firstRowFirstColumn bigFont paddingBottom" data-offset-x="0" data-offset-y="97" data-expected-width="100" data-expected-height="75">É</div>
  <div class="firstRowSecondColumn verticalLR" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="137">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 height25" data-offset-x="0" data-offset-y="137" data-expected-width="200" data-expected-height="25"></div>
</div>
offsetTop expected 97 but got 0
FAIL .inline-grid 2 assert_equals:
<div class="inline-grid alignItemsBaseline columns min-content-rows">
  <div class="firstRowFirstColumn bigFont paddingBottom" data-offset-x="0" data-offset-y="59" data-expected-width="100" data-expected-height="75">É</div>
  <div class="firstRowSecondColumn verticalLR" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="99">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 height25" data-offset-x="0" data-offset-y="99" data-expected-width="200" data-expected-height="25"></div>
</div>
offsetTop expected 59 but got 0
FAIL .inline-grid 3 assert_equals:
<div class="inline-grid alignItemsBaseline columns max-content-rows">
  <div class="firstRowFirstColumn bigFont paddingBottom" data-offset-x="0" data-offset-y="59" data-expected-width="100" data-expected-height="75">É</div>
  <div class="firstRowSecondColumn verticalLR" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="99">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 height25" data-offset-x="0" data-offset-y="99" data-expected-width="200" data-expected-height="25"></div>
</div>
offsetTop expected 59 but got 0
FAIL .inline-grid 4 assert_equals:
<div class="inline-grid alignItemsBaseline columns fit-content-rows">
  <div class="firstRowFirstColumn bigFont paddingBottom" data-offset-x="0" data-offset-y="59" data-expected-width="100" data-expected-height="75">É</div>
  <div class="firstRowSecondColumn verticalLR" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="99">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 height25" data-offset-x="0" data-offset-y="99" data-expected-width="200" data-expected-height="25"></div>
</div>
offsetTop expected 59 but got 0
FAIL .inline-grid 5 assert_equals:
<div class="inline-grid alignItemsBaseline columns flex-rows">
  <div class="firstRowFirstColumn bigFont paddingBottom" data-offset-x="0" data-offset-y="59" data-expected-width="100" data-expected-height="75">É</div>
  <div class="firstRowSecondColumn verticalLR" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="99">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 height25" data-offset-x="0" data-offset-y="99" data-expected-width="200" data-expected-height="25"></div>
</div>
offsetTop expected 59 but got 0
FAIL .inline-grid 6 assert_equals:
<div class="inline-grid alignItemsBaseline columns auto-rows">
  <div class="firstRowFirstColumn bigFont paddingBottom" data-offset-x="0" data-offset-y="59" data-expected-width="100" data-expected-height="75">É</div>
  <div class="firstRowSecondColumn verticalLR" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="99">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 height25" data-offset-x="0" data-offset-y="99" data-expected-width="200" data-expected-height="25"></div>
</div>
offsetTop expected 59 but got 0
FAIL .inline-grid 7 assert_equals:
<div class="inline-grid justifyItemsBaseline rows width200">
  <div class="firstRowFirstColumn bigFont verticalLR paddingLeft" data-offset-x="0" data-offset-y="0" data-expected-width="75" data-expected-height="100">É</div>
  <div class="secondRowFirstColumn" data-offset-x="35" data-offset-y="100" data-expected-width="175" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="firstRowSpanning2AutoColumn width25" data-offset-x="175" data-offset-y="0" data-expected-width="25" data-expected-height="200"></div>
</div>
offsetLeft expected 35 but got 0
FAIL .inline-grid 8 assert_equals:
<div class="inline-grid justifyItemsBaseline rows min-content-columns">
  <div class="firstRowFirstColumn bigFont verticalLR paddingLeft" data-offset-x="0" data-offset-y="0" data-expected-width="75" data-expected-height="100">É</div>
  <div class="secondRowFirstColumn" data-offset-x="35" data-offset-y="100" data-expected-width="99" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="firstRowSpanning2AutoColumn width25" data-offset-x="99" data-offset-y="0" data-expected-width="25" data-expected-height="200"></div>
</div>
width expected 99 but got 75
FAIL .inline-grid 9 assert_equals:
<div class="inline-grid justifyItemsBaseline rows max-content-columns">
  <div class="firstRowFirstColumn bigFont verticalLR paddingLeft" data-offset-x="0" data-offset-y="0" data-expected-width="75" data-expected-height="100">É</div>
  <div class="secondRowFirstColumn" data-offset-x="35" data-offset-y="100" data-expected-width="416" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="firstRowSpanning2AutoColumn width25" data-offset-x="451" data-offset-y="0" data-expected-width="25" data-expected-height="200"></div>
</div>
offsetLeft expected 35 but got 0
FAIL .inline-grid 10 assert_equals:
<div class="inline-grid justifyItemsBaseline rows fit-content-columns">
  <div class="firstRowFirstColumn bigFont verticalLR paddingLeft" data-offset-x="0" data-offset-y="0" data-expected-width="75" data-expected-height="100">É</div>
  <div class="secondRowFirstColumn" data-offset-x="35" data-offset-y="100" data-expected-width="416" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="firstRowSpanning2AutoColumn width25" data-offset-x="451" data-offset-y="0" data-expected-width="25" data-expected-height="200"></div>
</div>
offsetLeft expected 35 but got 0
FAIL .inline-grid 11 assert_equals:
<div class="inline-grid justifyItemsBaseline rows flex-columns">
  <div class="firstRowFirstColumn bigFont verticalLR paddingLeft" data-offset-x="0" data-offset-y="0" data-expected-width="75" data-expected-height="100">É</div>
  <div class="secondRowFirstColumn" data-offset-x="35" data-offset-y="100" data-expected-width="416" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="firstRowSpanning2AutoColumn width25" data-offset-x="451" data-offset-y="0" data-expected-width="25" data-expected-height="200"></div>
</div>
offsetLeft expected 35 but got 0
FAIL .inline-grid 12 assert_equals:
<div class="inline-grid justifyItemsBaseline auto-rows max-flex-columns">
  <div class="firstRowFirstColumn bigFont verticalLR paddingLeft" data-offset-x="0" data-offset-y="0" data-expected-width="75" data-expected-height="50">É</div>
  <div class="secondRowFirstColumn verticalLR" data-offset-x="32" data-offset-y="50" data-expected-width="80" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="firstRowSpanning2AutoColumn width25" data-offset-x="112" data-offset-y="0" data-expected-width="25" data-expected-height="150"></div>
</div>
width expected 80 but got 85
FAIL .inline-grid 13 assert_equals:
<div class="inline-grid verticalLR alignItemsBaseline columns width200">
  <div class="firstRowFirstColumn bigFont paddingLeft" data-offset-x="0" data-offset-y="0" data-expected-width="75" data-expected-height="100">É</div>
  <div class="firstRowSecondColumn horizontalTB" data-offset-x="35" data-offset-y="100" data-expected-width="175" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="175" data-offset-y="0" data-expected-width="25" data-expected-height="200"></div>
</div>
offsetLeft expected 35 but got 0
FAIL .inline-grid 14 assert_equals:
<div class="inline-grid verticalLR alignItemsBaseline columns min-content-rows">
  <div class="firstRowFirstColumn bigFont paddingLeft" data-offset-x="0" data-offset-y="0" data-expected-width="75" data-expected-height="100">É</div>
  <div class="firstRowSecondColumn horizontalTB" data-offset-x="35" data-offset-y="100" data-expected-width="99" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="99" data-offset-y="0" data-expected-width="25" data-expected-height="200"></div>
</div>
width expected 99 but got 75
FAIL .inline-grid 15 assert_equals:
<div class="inline-grid verticalLR alignItemsBaseline columns max-content-rows">
  <div class="firstRowFirstColumn bigFont paddingLeft" data-offset-x="0" data-offset-y="0" data-expected-width="75" data-expected-height="100">É</div>
  <div class="firstRowSecondColumn horizontalTB" data-offset-x="35" data-offset-y="100" data-expected-width="416" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="451" data-offset-y="0" data-expected-width="25" data-expected-height="200"></div>
</div>
offsetLeft expected 35 but got 0
FAIL .inline-grid 16 assert_equals:
<div class="inline-grid verticalLR alignItemsBaseline columns fit-content-rows">
  <div class="firstRowFirstColumn bigFont paddingLeft" data-offset-x="0" data-offset-y="0" data-expected-width="75" data-expected-height="100">É</div>
  <div class="firstRowSecondColumn horizontalTB" data-offset-x="35" data-offset-y="100" data-expected-width="416" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="451" data-offset-y="0" data-expected-width="25" data-expected-height="200"></div>
</div>
offsetLeft expected 35 but got 0
FAIL .inline-grid 17 assert_equals:
<div class="inline-grid verticalLR alignItemsBaseline columns flex-rows">
  <div class="firstRowFirstColumn bigFont paddingLeft" data-offset-x="0" data-offset-y="0" data-expected-width="75" data-expected-height="100">É</div>
  <div class="firstRowSecondColumn horizontalTB" data-offset-x="35" data-offset-y="100" data-expected-width="416" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="451" data-offset-y="0" data-expected-width="25" data-expected-height="200"></div>
</div>
offsetLeft expected 35 but got 0
FAIL .inline-grid 18 assert_equals:
<div class="inline-grid verticalLR alignItemsBaseline auto-columns max-flex-rows">
  <div class="firstRowFirstColumn bigFont" data-offset-x="0" data-offset-y="0" data-expected-width="50" data-expected-height="50">É</div>
  <div class="firstRowSecondColumn horizontalBT" data-offset-x="7" data-offset-y="50" data-expected-width="80" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="87" data-offset-y="0" data-expected-width="25" data-expected-height="150"></div>
</div>
width expected 80 but got 85
FAIL .inline-grid 19 assert_equals:
<div class="inline-grid verticalRL alignItemsBaseline columns width200">
  <div class="firstRowFirstColumn bigFont paddingLeft" data-offset-x="28" data-offset-y="0" data-expected-width="75" data-expected-height="100">É</div>
  <div class="firstRowSecondColumn horizontalTB" data-offset-x="63" data-offset-y="100" data-expected-width="137" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="38" data-offset-y="0" data-expected-width="25" data-expected-height="200"></div>
</div>
offsetLeft expected 28 but got 125
FAIL .inline-grid 20 assert_equals:
<div class="inline-grid verticalRL alignItemsBaseline columns min-content-rows">
  <div class="firstRowFirstColumn bigFont paddingLeft" data-offset-x="-10" data-offset-y="0" data-expected-width="75" data-expected-height="100">É</div>
  <div class="firstRowSecondColumn horizontalTB" data-offset-x="25" data-offset-y="100" data-expected-width="99" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="0" data-offset-y="0" data-expected-width="25" data-expected-height="200"></div>
</div>
offsetLeft expected -10 but got 25
FAIL .inline-grid 21 assert_equals:
<div class="inline-grid verticalRL alignItemsBaseline columns max-content-rows">
  <div class="firstRowFirstColumn bigFont paddingLeft" data-offset-x="-10" data-offset-y="0" data-expected-width="75" data-expected-height="100">É</div>
  <div class="firstRowSecondColumn horizontalTB" data-offset-x="25" data-offset-y="100" data-expected-width="99" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="0" data-offset-y="0" data-expected-width="25" data-expected-height="200"></div>
</div>
offsetLeft expected -10 but got 366
FAIL .inline-grid 22 assert_equals:
<div class="inline-grid verticalRL alignItemsBaseline auto-columns fit-content-rows">
  <div class="firstRowFirstColumn bigFont paddingLeft" data-offset-x="-10" data-offset-y="0" data-expected-width="75" data-expected-height="50">É</div>
  <div class="firstRowSecondColumn horizontalTB" data-offset-x="25" data-offset-y="50" data-expected-width="99" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="0" data-offset-y="0" data-expected-width="25" data-expected-height="150"></div>
</div>
offsetLeft expected -10 but got 366
FAIL .inline-grid 23 assert_equals:
<div class="inline-grid verticalRL alignItemsBaseline auto-columns flex-rows">
  <div class="firstRowFirstColumn bigFont paddingLeft" data-offset-x="-10" data-offset-y="0" data-expected-width="75" data-expected-height="50">É</div>
  <div class="firstRowSecondColumn horizontalTB" data-offset-x="25" data-offset-y="50" data-expected-width="99" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="0" data-offset-y="0" data-expected-width="25" data-expected-height="150"></div>
</div>
offsetLeft expected -10 but got 366
FAIL .inline-grid 24 assert_equals:
<div class="inline-grid verticalRL alignItemsBaseline auto-columns max-flex-rows">
  <div class="firstRowFirstColumn bigFont paddingLeft" data-offset-x="-10" data-offset-y="0" data-expected-width="75" data-expected-height="50">É</div>
  <div class="firstRowSecondColumn horizontalTB" data-offset-x="25" data-offset-y="50" data-expected-width="99" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="0" data-offset-y="0" data-expected-width="25" data-expected-height="150"></div>
</div>
width expected 99 but got 416
